physio-web/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
pyttsx3==2.90
gtts==2.4.0
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.1.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
h5py==3.15.1
